Two pointer solution to Remove Duplicates from Sorted Array

+++ problems/0027-RemoveElement/src/solution_default.cpp
@@ -11,6 +11,6 @@ int SolutionDefault::removeElement(std::vector<int>& nums, int val) {
     }
 
     nums.resize(length);
-    return length;
+    return static_cast<int>(length);
 
 }
